% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{list_books}
\alias{list_books}
\title{List the Quarto books of a folder}
\usage{
list_books(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{Folder to search in.}

\item{recursive}{Whether to search in subfolders.}
}
\value{
The names (relative to \code{path}) of the folders containing a
\verb{_quarto.yml} of type \code{book}.
}
\description{
List the Quarto books of a folder
}
\examples{
tmp <- tempfile()
dir.create(tmp)
create_book(c("rapport_a", "rapport_b"), path = tmp)
list_books(tmp)
}
