% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{remove_template_book}
\alias{remove_template_book}
\title{Remove a user book template}
\usage{
remove_template_book(name)
}
\arguments{
\item{name}{Template name.}
}
\value{
\code{TRUE} invisibly if the template was removed.
}
\description{
Remove a user book template
}
