% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_dir}
\alias{template_dir}
\title{Folder where user templates are stored}
\usage{
template_dir()
}
\value{
A path.
}
\description{
Returns the option \code{kquarto.r.template_dir} if set, otherwise
\code{tools::R_user_dir("kquarto.r", "data")/templates}.
}
\examples{
template_dir()
}
